+++ CHANGELOG.md
@@ -6,8 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.6.3] - 2025-12-17
 ### Fixed
 - Spectators seeing vanished player waypoints
+- Display name cache causing issues with some datapacks
 
 ## [1.6.2] - 2025-11-23
 ### Fixed